requirements.txt
django~=3.1.7
django-crispy-forms~=1.12.0
geopy~=2.2.0
geoip2~=4.2.0
folium~=0.12.1
django-bootstrap4~=3.0.1
django-bootstrap-modal-forms~=2.2.0
django-bootstrap-datepicker-plus~=3.0.5
pillow~=8.1.0
